Add files via upload
Overview : This repository contains a comprehensive HR Analytics project focused on building a predictive model to identify potential candidates for promotion within an organization. The project leverages data analysis and machine learning techniques to provide HR professionals with a tool that aids in making data-driven decisions related to employee promotions.
Objective : The HR Analytics Promotion Model is designed to streamline and optimize the process of identifying and promoting high-potential employees. By analyzing historical data and relevant features, the model can predict which employees are most likely to excel in higher roles, ensuring a fair and data-driven promotion process.